GDAX orderbook db
Not sure if this is a functional commit, but on my system I have it working in a different directory. Don't pip install gdax, although you may need to install the dependencies. The pip install version of gdax seemed to be broken. If you get pull their repo it works.